Support UTF-8 encoded file paths on Windows
This assumes the file paths will be UTF-8 encoded, they most likely
will be with recent versions of R. The paths are converted to UTF-16 by
`MultiByteToWideChar()` before being passed to `_wfopen()`.

On other systems we call `fopen()` directly, as the code did before.

I also pulled out the individual calls to `fopen()` and put them above
the switch to reduce some repetition in the code.

There is a also a simple test which demonstrates the issue.

Fixes #89